+++ src/features/task-detail/hooks/usePdfPinInteraction.ts
@@ -1,0 +1,50 @@
+import { useRef } from 'react';
+import type { PinWithAuthor } from '@/features/task-detail/types/taskDetailType';
+import { useTaskDetailStore } from '@/features/task-detail/store/useTaskDetailStore';
+import type { PageSize } from '@/features/task-detail/types/pdfTypes'; // width, height 타입
+import { useAuthStore } from '@/features/auth/store/authStore';
+
+export const usePdfPinInteraction = (pageNumber: number, pageSize: PageSize) => {
+  const mouseMoved = useRef(false);
+  const { user } = useAuthStore();
+  const { currentPin, setCurrentPin, selectedFile } = useTaskDetailStore();
+
+  const handleOverlayClick = (e: React.MouseEvent) => {
+    if (mouseMoved.current) return;
+
+    const rect = (e.currentTarget as HTMLDivElement).getBoundingClientRect();
+    const x = e.clientX - rect.left;
+    const y = e.clientY - rect.top;
+
+    const pdfX = (x / rect.width) * pageSize.width;
+    const pdfY = (1 - y / rect.height) * pageSize.height;
+
+    if (
+      currentPin &&
+      Math.abs(currentPin.fileX! - pdfX) < 20 &&
+      Math.abs(currentPin.fileY! - pdfY) < 20 &&
+      currentPin.filePage === pageNumber
+    ) {
+      setCurrentPin(null);
+      return;
+    }
+
+    const newPin: PinWithAuthor = {
+      fileId: selectedFile?.fileId ?? '',
+      fileName: selectedFile?.fileName ?? '',
+      filePage: pageNumber,
+      fileX: pdfX,
+      fileY: pdfY,
+      author: {
+        id: user?.id ?? '',
+        name: user?.name ?? '익명',
+        avatar: user?.avatar ?? '0',
+        backgroundColor: user?.backgroundColor ?? '#CCCCCC',
+      },
+    };
+
+    setCurrentPin(newPin);
+  };
+
+  return { handleOverlayClick, mouseMoved };
+};
